Alexa vegetarian restaurants
Amazon's Alexa team is working on optimizing the customer experience for scenarios where customers ask generic questions. One example of a generic question is 'What are good vegetarian restaurants nearby?'. In this scenario, Alexa would then search for a list of vegetarian restaurants in the city and select the nearest X vegatarian restaurants relative to the customer's location.

Given an array representing the locations of N vegetarian restaurants in the city, this algorithm finds the nearest X vegetarian restaurants to the customer's location.