Building lot
You are in charge of preparing a recently purchased lot for one of Amazon's new buildings. The lot is covered with trenches and has a single obstacle that needs to be taken down before the foundation can be prepared for the building. The demolition robot must remove the obstacle before process can be made on the building.

The algorithm implemented determines the minimum distance required for the demolition robot to remove the obstacle.